Add windows named pipe listener to engine-session.
